app/requirements.txt
joblib==1.4.2
pandas==2.2.3
sklearn-compat==0.1.3
streamlit==1.42.0
